app/ml_model.py
import joblib
import pandas as pd
from config.settings import Config

class VulnScoringModel:
    def __init__(self):
        self.model = joblib.load(Config.VULN_MODEL_PATH)

    def score_vulnerabilities(self, scan_data: pd.DataFrame) -> pd.DataFrame:
        # Add ML scoring logic here
        scan_data['risk_score'] = self.model.predict(scan_data[['severity', 'cvss']])
        return scan_data
