requirements.txt
streamlit
openai
python-dotenv
pandas
scikit-learn
numpy
requests
plotly